#5 feat: onload 시 또는 버튼 클릭시 현재 좌표로 이동 함수 추가
