+++ front/src/apis/useGetMyLoc.js
@@ -1,0 +1,65 @@
+const { kakao } = window;
+
+export function useGetMyLoc() {
+  const container = document.getElementById('Map'); // 가이드는 Map이다
+  const options = {
+    center: new kakao.maps.LatLng(33.450701, 126.570667),
+    level: 2,
+  };
+  const map = new kakao.maps.Map(container, options);
+
+  // HTML5의 geolocation으로 사용할 수 있는지 확인합니다
+  if (navigator.geolocation) {
+    // GeoLocation을 이용해서 접속 위치를 얻어옵니다
+    navigator.geolocation.getCurrentPosition(function (position) {
+      var lat = position.coords.latitude, // 위도
+        lon = position.coords.longitude; // 경도
+
+      var locPosition = new kakao.maps.LatLng(lat, lon), // 마커가 표시될 위치를 geolocation으로 얻어온 좌표로 생성합니다
+        message = '<div style="padding:5px;">여기에 계신가요?!</div>'; // 인포윈도우에 표시될 내용입니다
+
+      // 마커와 인포윈도우를 표시합니다
+      displayMarker(locPosition, message);
+    });
+  } else {
+    // HTML5의 GeoLocation을 사용할 수 없을때 마커 표시 위치와 인포윈도우 내용을 설정합니다
+
+    var locPosition = new kakao.maps.LatLng(33.450701, 126.570667),
+      message = 'geolocation을 사용할수 없어요..';
+
+    displayMarker(locPosition, message);
+  }
+
+  // 마커 이미지를 가져옵니다.
+  var imageSrc = 'http://getdrawings.com/free-icon/coke-icon-70.png', // 마커이미지의 주소입니다
+    imageSize = new kakao.maps.Size(60, 60), // 마커이미지의 크기입니다
+    imageOption = { offset: new kakao.maps.Point(27, 69) }; // 마커이미지의 옵션입니다. 마커의 좌표와 일치시킬 이미지 안에서의 좌표를 설정합니다.
+
+  var markerImage = new kakao.maps.MarkerImage(imageSrc, imageSize, imageOption);
+  // 지도에 마커와 인포윈도우를 표시하는 함수입니다
+
+  function displayMarker(locPosition, message) {
+    // 마커를 생성합니다
+    var marker = new kakao.maps.Marker({
+      map: map,
+      position: locPosition,
+      image: markerImage,
+      // draggable: true, // 마커를 드래그 가능하도록 설정한다
+    });
+
+    var iwContent = message, // 인포윈도우에 표시할 내용
+      iwRemoveable = true;
+
+    // 인포윈도우를 생성합니다
+    var infowindow = new kakao.maps.InfoWindow({
+      content: iwContent,
+      removable: iwRemoveable,
+    });
+
+    // 인포윈도우를 마커위에 표시합니다
+    infowindow.open(map, marker);
+
+    // 지도 중심좌표를 접속위치로 변경합니다
+    map.setCenter(locPosition);
+  }
+}
